+++ packages/node/src/sync/stream/NodeStreamingSyncImplementation.ts
@@ -4,18 +4,18 @@ import {
   LockOptions,
   LockType
 } from '@powersync/common';
-import Lock from 'async-lock';
+import { Mutex } from 'async-mutex';
 
 /**
  * Global locks which prevent multiple instances from syncing
  * concurrently.
  */
-const LOCKS = new Map<string, Lock>();
+const LOCKS = new Map<string, Map<LockType, Mutex>>();
 
 const lockTypes = new Set(Object.values(LockType));
 
 export class NodeStreamingSyncImplementation extends AbstractStreamingSyncImplementation {
-  locks: Lock;
+  locks: Map<LockType, Mutex>;
 
   constructor(options: AbstractStreamingSyncImplementationOptions) {
     super(options);
@@ -32,7 +32,9 @@ export class NodeStreamingSyncImplementation extends AbstractStreamingSyncImplem
       return;
     }
 
-    this.locks = new Lock();
+    this.locks = new Map<LockType, Mutex>();
+    this.locks.set(LockType.CRUD, new Mutex());
+    this.locks.set(LockType.SYNC, new Mutex());
 
     if (identifier) {
       LOCKS.set(identifier, this.locks);
@@ -40,10 +42,11 @@ export class NodeStreamingSyncImplementation extends AbstractStreamingSyncImplem
   }
 
   obtainLock<T>(lockOptions: LockOptions<T>): Promise<T> {
-    if (!lockTypes.has(lockOptions.type)) {
+    const lock = this.locks.get(lockOptions.type);
+    if (!lock) {
       throw new Error(`Lock type ${lockOptions.type} not found`);
     }
-    return this.locks.acquire(lockOptions.type, async () => {
+    return lock.runExclusive(async () => {
       if (lockOptions.signal?.aborted) {
         throw new Error('Aborted');
       }