+++ packages/adapter-sql-js/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @powersync/adapter-sql-js
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [66218b2]
+- Updated dependencies [3af4a2c]
+  - @powersync/common@1.42.0
+
+## 0.0.10
+
+### Patch Changes
+
+- 3e4a25c: Don't minify releases, enable source maps.
+- Updated dependencies [3e4a25c]
+  - @powersync/common@1.41.1
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [2f8b30c]
+  - @powersync/common@1.41.0
+
 ## 0.0.8
 
 ### Patch Changes