+++ packages/tanstack-react-query/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @powersync/tanstack-react-query
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [66218b2]
+- Updated dependencies [3af4a2c]
+  - @powersync/common@1.42.0
+  - @powersync/react@1.8.2
+
+## 0.1.8
+
+### Patch Changes
+
+- 3e4a25c: Don't minify releases, enable source maps.
+- Updated dependencies [3e4a25c]
+  - @powersync/common@1.41.1
+  - @powersync/react@1.8.2
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [2f8b30c]
+  - @powersync/common@1.41.0
+  - @powersync/react@1.8.1
+
 ## 0.1.6
 
 ### Patch Changes