+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @powersync/vue
 
+## 0.4.1
+
+### Patch Changes
+
+- 3e4a25c: Don't minify releases, enable source maps.
+- Updated dependencies [3e4a25c]
+  - @powersync/common@1.41.1
+
 ## 0.4.0
 
 ### Minor Changes